+++ canaille/ldap_backend/backend.py
@@ -22,7 +22,9 @@ def setup_ldap_models(config):
     User.object_class = [config["LDAP"].get("USER_CLASS", User.DEFAULT_OBJECT_CLASS)]
 
     group_base = (
-        config["LDAP"].get("GROUP_BASE", "").replace(f',{config["LDAP"]["ROOT_DN"]}', "")
+        config["LDAP"]
+        .get("GROUP_BASE", "")
+        .replace(f',{config["LDAP"]["ROOT_DN"]}', "")
     )
     Group.base = group_base or None
     Group.id = config["LDAP"].get("GROUP_ID_ATTRIBUTE", Group.DEFAULT_ID_ATTRIBUTE)